re upload

+++ app/Http/Controllers/Api/JobController.php
@@ -15,27 +15,27 @@ class JobController extends BaseController
 {
     public function getJobList(Request $request)
     {
-            $userType = $request->attributes->get('type');
+        $userType = $request->attributes->get('type');
 
-            $loggedUserId = $this->getLoggedUserId(); // Assuming you have a method to get the logged user's ID
-            
-            if ($userType === 'admin' || $userType === 'recruiter') {
-                $jobs = DB::table('jobs as j')
-                    ->select('j.*', 's.name as student_name', 'a.student_id', 'r.name as recruiter_name', 't.title as test_name')
-                    ->leftJoin('job_applications as a', 'a.job_id', '=', 'j.id')
-                    ->leftJoin('students as s', 's.auth_id', '=', 'a.student_id')
-                    ->leftJoin('recruiters as r', 'r.auth_id', '=', 'j.recruiter_id')
-                    ->leftJoin('job_tests as t', 't.id', '=', 'j.test_id');
-            
-                if ($userType === 'recruiter') {
-                    $jobs->where('j.recruiter_id', '=', $loggedUserId);
-                }
+        $loggedUserId = $this->getLoggedUserId(); // Assuming you have a method to get the logged user's ID
 
-                // Log::info("Getting jobs query: ", ['query' => $jobs->toSql(), 'bindings' => $jobs->getBindings()]);
-                $jobs= $jobs->get();
-    // Log::info("getting jobs", $jobs);
-                return $this->sendResponse(['jobs' => $jobs]);
+        if ($userType === 'admin' || $userType === 'recruiter') {
+            $jobs = DB::table('jobs as j')
+                ->select('j.*', 's.name as student_name', 'a.student_id', 'r.name as recruiter_name', 't.title as test_name', 't.instruction as test_instruction')
+                ->leftJoin('job_applications as a', 'a.job_id', '=', 'j.id')
+                ->leftJoin('students as s', 's.auth_id', '=', 'a.student_id')
+                ->leftJoin('recruiters as r', 'r.auth_id', '=', 'j.recruiter_id')
+                ->leftJoin('job_tests as t', 't.id', '=', 'j.test_id');
+
+            if ($userType === 'recruiter') {
+                $jobs->where('j.recruiter_id', '=', $loggedUserId);
             }
+
+            // Log::info("Getting jobs query: ", ['query' => $jobs->toSql(), 'bindings' => $jobs->getBindings()]);
+            $jobs = $jobs->get();
+            // Log::info("getting jobs", $jobs);
+            return $this->sendResponse(['jobs' => $jobs]);
+        }
 
         if ($userType === 'student') {
 
@@ -48,9 +48,11 @@ class JobController extends BaseController
             $student_class_id = $student->class_id;
 
             $jobs = DB::table('jobs as j')
-                ->select('j.*', 'r.name as recruiter_name')
+                ->select('j.*', 'r.name as recruiter_name','t.instruction as test_instruction')
                 ->leftJoin('recruiters as r', 'r.auth_id', '=', 'j.recruiter_id')
+                ->leftJoin('job_tests as t', 't.id', '=', 'j.test_id')
                 ->whereRaw("FIND_IN_SET(?, j.class_id)", [$student_class_id])
+                ->where('status', true)
                 ->get();
 
             foreach ($jobs as $job) {
@@ -76,6 +78,7 @@ class JobController extends BaseController
 
     public function storeJobDetails(Request $request)
     {
+        log::info("Store", $request->all());
         $validator = Validator::make($request->all(), [
             'title' => 'required|string|max:255',
             'image' => 'required',
@@ -83,22 +86,22 @@ class JobController extends BaseController
             'location' => 'required',
             'criteria' => 'required',
             'description' => 'required',
-            'test_id' => 'required',
+            // 'test_id' => 'nullable|exists:job_tests,id',
             'recruiter_id' => 'required',
             'instruction' => 'required',
             'passing_percentage' => [
-                'required_if:test_id,!=,0',
+                'nullable',
                 'numeric',
                 'min:0',
                 function ($attribute, $value, $fail) use ($request) {
-                    if ($request->test_id != 0 && $value >= 100) {
+                    if ($request->test_id && $value >= 100) {
                         $fail($attribute . ' must be less than 100.');
                     }
                 },
             ],
             'selectedClass' => 'required',
         ]);
-    
+
 
         if ($validator->fails()) {
             return $this->sendValidationError($validator);
@@ -120,13 +123,15 @@ class JobController extends BaseController
             $job->location = $request->location;
             $job->criteria = $request->criteria;
             $job->description = $request->description;
-            $job->test_id = $request->test_id;
+            $job->test_id = $request->test_id ? $request->test_id : null;
             $job->recruiter_id = $request->recruiter_id;
-            if ($request->test_id == 0) {
-                $job->passing_percentage = 0;
-            } else {
-                $job->passing_percentage = $request->passing_percentage;
-            }
+            // if ($request->test_id == 0) {
+            //     $job->passing_percentage = 0;
+            // } else {
+            //     $job->passing_percentage = $request->passing_percentage;
+            // }
+
+            $job->passing_percentage = $request->test_id ? $request->passing_percentage : 0;
             $job->instruction = $request->instruction;
             $job->status = 1;
             if ($job->save()) {
@@ -139,26 +144,28 @@ class JobController extends BaseController
 
     public function updateJobDetails(Request $request, $jobId)
     {
-        // Log::info("updateing job0", $request->all());
+        Log::info("updating job", $request->all());
         $validator = Validator::make($request->all(), [
             'title' => 'required|string|max:255',
             'annual_ctc' => 'required',
             'location' => 'required',
             'criteria' => 'required',
             'description' => 'required',
-            'test_id' => 'required',
+            // 'test_id' => 'nullable|exists:job_tests,id',
             'recruiter_id' => 'required',
             'instruction' => 'required',
+            'status' => 'required',
             'passing_percentage' => [
-                'required',
+                'nullable',
                 'numeric',
                 'min:0',
-                function ($attribute, $value, $fail) {
-                    if ($value >= 100) {
-                        $fail('Passing Percentage must be less than 100.');
+                function ($attribute, $value, $fail) use ($request) {
+                    if ($request->test_id && $value >= 100) {
+                        $fail($attribute . ' must be less than 100.');
                     }
                 },
             ],
+            'selectedClass' => 'required',
         ]);
 
         if ($validator->fails()) {
@@ -181,11 +188,18 @@ class JobController extends BaseController
             $job->annual_ctc = $request->annual_ctc;
             $job->location = $request->location;
             $job->criteria = $request->criteria;
+            $job->status = $request->status;
             $job->description = $request->description;
-            $job->test_id = $request->test_id;
+            // $job->test_id = $request->test_id ?: null;
+
+            // $job->test_id = $request->test_id ?: null;
+            $job->test_id = $request->test_id ? $request->test_id : null;
+
             $job->recruiter_id = $request->recruiter_id;
-            $job->passing_percentage = $request->passing_percentage;
+            $job->passing_percentage = $request->test_id ? $request->passing_percentage : 0;
             $job->instruction = $request->instruction;
+            $job->class_Id = $request->selectedClass;
+
             if ($job->save()) {
                 return $this->sendResponse([], 'Job updated successfully.');
             } else {
@@ -194,60 +208,89 @@ class JobController extends BaseController
         }
     }
 
+
+
     public function deleteJobDetails(Request $request, $jobId)
     {
-
+        // Find the job
         $job = Job::find($jobId);
-
+    
+        if (!$job) {
+            return response()->json(['message' => 'Job not found'], 404);
+        }
+    
+        // Delete job applications associated with this job
+        JobApplication::where('job_id', $jobId)->delete();
+    
+        // Delete the job
         $job->delete();
-
+    
         return $this->sendResponse([], 'Job deleted successfully');
     }
-
+    
     public function applyJob(Request $request)
     {
+        Log::info("requests applyjob", $request->all());
+
+        // Validate with conditional rule for testId
         $validator = Validator::make($request->all(), [
             'jobId' => 'required',
             'studentId' => 'required',
             'schoolId' => 'required',
             'classId' => 'required',
-            'testId' => 'required',
+            'testId' => 'nullable', // testId is optional
         ]);
-    
+
         if ($validator->fails()) {
-            return $this->sendValidationError($validator);
-        } else {
-            $data = $validator->validated();
-    
-            $existingSession = JobApplication::where('student_id', $data['studentId'])
-                ->where('job_id', $data['jobId'])
-                ->first();
-    
-            if ($existingSession) {
-                return response()->json(['message' => 'Test already taken'], 409); // 409 Conflict
-            }
-    
-            $token = Str::random(32); 
+            return response()->json(['message' => $validator->errors()], 400);
+        }
+
+        $data = $validator->validated();
+
+        // Check if the job application already exists
+        $existingSession = JobApplication::where('student_id', $data['studentId'])
+            ->where('job_id', $data['jobId'])
+            ->first();
+
+        if ($existingSession) {
+            return response()->json(['message' => 'Test already taken'], 409); // 409 Conflict
+        }
+
+        if (isset($data['testId'])) {
+            $token = Str::random(32);
+
             // Create a new test session
             $testSession = JobApplication::create([
                 'student_id' => $data['studentId'],
                 'school_id' => $data['schoolId'],
-                'test_id' => $data['testId'], // assuming 'latestTestId' should be 'testId'
-                'class_id' => $data['classId'], // assuming 'latestTestId' should be 'testId'
-                'job_id' => $data['jobId'], // assuming 'latestTestId' should be 'testId'
-                // Set additional fields as necessary
+                'test_id' => $data['testId'],
+                'class_id' => $data['classId'],
+                'job_id' => $data['jobId'],
                 'token' => $token,
             ]);
-    
-            return $this->sendResponse(['token' => $token, 'status' => 200], 200);
+
+            // return response()->json(['token' => $token, 'status' => 200], 200);
+            return $this->sendResponse(['token' => $token, 'status' => 200], 'Job updated successfully.');
+        } else {
+            // Create a job application without test
+            $jobApplication = JobApplication::create([
+                'student_id' => $data['studentId'],
+                'school_id' => $data['schoolId'],
+                'class_id' => $data['classId'],
+                'job_id' => $data['jobId'],
+                'is_completed' => true,
+                'is_pass' => true,
+                'is_test' => false,
+            ]);
+            return $this->sendResponse(['status' => 200, 'message' => 'Job application submitted successfully'], 'Job updated successfully.');
+
+            // return response()->json(['status' => 200, 'message' => 'Job application submitted successfully'], 200);
         }
     }
-    
 
 
     public function getStudentJobApplications(Request $request, $jobId)
     {
-
         $job_applications = DB::table('job_applications as a')
             ->select('j.*', 's.name as student_name', 'a.student_id', 'c.name as class_name', 'a.is_pass')
             ->leftJoin('jobs as j', 'j.id', '=', 'a.job_id')
@@ -279,47 +322,46 @@ class JobController extends BaseController
     }
 
     public function getJobTestDetailsByToken(Request $request, $token, $jobId)
-{
-    // Find the test result entry using the token
-    $testResult = DB::table('job_applications')
-                    ->where('token', $token)
-                    ->where('job_id', $jobId)
-                    ->first();
-    
-    $job = DB::table('jobs')
-                    ->where('id', $jobId)
-                    ->first();
-    
-    // Check if the test result does not exist, which indicates an invalid token or test ID
-    if (!$testResult) {
-        return $this->sendError("Test or token not found", [], 404); // Use HTTP 404 Not Found for resource not found
+    {
+        // Find the test result entry using the token
+        $testResult = DB::table('job_applications')
+            ->where('token', $token)
+            ->where('job_id', $jobId)
+            ->first();
+
+        $job = DB::table('jobs')
+            ->where('id', $jobId)
+            ->first();
+
+        // Check if the test result does not exist, which indicates an invalid token or test ID
+        if (!$testResult) {
+            return $this->sendError("Test or token not found", [], 404); // Use HTTP 404 Not Found for resource not found
+        }
+
+        // Check if the test has already been taken
+        if ($testResult->token_status == 1) {
+            return $this->sendResponse([], "Test has already been taken", false);
+        }
+
+        // Update the status to 1 to mark as taken
+        DB::table('job_applications')
+            ->where('id', $testResult->id)
+            ->update(['token_status' => 1]);
+
+        // Fetch the term test details
+        $job_test = DB::table('job_tests as a')
+            ->select('a.id',   'a.title',  'a.description', 'a.total_score', 'a.time_limit', 'a.no_of_questions',   'a.question_ids', 'a.passing_percentage')
+            ->where('a.id', $testResult->test_id)
+            ->first();
+
+        if ($job_test && $job_test->question_ids) {
+            $questionIds = explode(',', $job_test->question_ids);
+            $job_test->questions = DB::table('job_questions')
+                ->whereIn('id', $questionIds)
+                ->get();
+        }
+
+        // Return the test details along with the test_result_id
+        return $this->sendResponse(['job_test' => $job_test, 'test_result' => $testResult, 'job' => $job], "Test details retrieved successfully", true);
     }
-
-    // Check if the test has already been taken
-    if ($testResult->token_status == 1 ) {
-        return $this->sendResponse([], "Test has already been taken", false); 
-    }
-
-    // Update the status to 1 to mark as taken
-    DB::table('job_applications')
-        ->where('id', $testResult->id)
-        ->update(['token_status' => 1]);
-
-    // Fetch the term test details
-    $job_test = DB::table('job_tests as a')
-                    ->select('a.id',   'a.title',  'a.description', 'a.total_score', 'a.time_limit', 'a.no_of_questions',   'a.question_ids', 'a.passing_percentage')   
-                    ->where('a.id', $testResult->test_id)
-                    ->first();
-
-    if ($job_test && $job_test->question_ids) {
-        $questionIds = explode(',', $job_test->question_ids);
-        $job_test->questions = DB::table('job_questions')
-                                    ->whereIn('id', $questionIds)
-                                    ->get();
-    }
-
-    // Return the test details along with the test_result_id
-    return $this->sendResponse(['job_test' => $job_test,'test_result'=>$testResult,'job'=> $job], "Test details retrieved successfully", true);
-}
-
 }
